requirements.txt
fabric2==3.2.2
pyyaml==6.0.2